move util functions to new packages
