+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/app/navDto/serializableNavType.kt
@@ -1,4 +1,4 @@
-package pw.checkers.navigation
+package pw.checkers.app.navDto
 
 import androidx.core.bundle.Bundle
 import androidx.core.uri.UriUtils
