+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/core/util/Constants.kt
@@ -1,4 +1,4 @@
-package pw.checkers.util
+package pw.checkers.core.util
 
 object Constants {
     const val SERVER_ADDRESS = "ws://localhost:8080/ws"
