+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/core/util/DoNothing.kt
@@ -1,0 +1,3 @@
+package pw.checkers.core.util
+
+typealias DoNothing = Unit
