+++ frontend/composeApp/src/wasmJsMain/kotlin/pw/checkers/main.kt
@@ -3,6 +3,7 @@ package pw.checkers
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
 import kotlinx.browser.document
+import pw.checkers.app.App
 
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {
